Implementation of stratified split of an InstanceList.
The existing split() method of an instance list only does random
splitting of the data. Stratified sampling is known to give better
results when it comes to cross validation. The current implementation
extends the InstanceList functionality with a stratified split of the
data, for as many target classes as possible.